variable stream window size
The stream window now starts at 64k and doubles after 64kb have been read by the
application, then doubles again after 128k total have been read, and so on up
to a maximum window size of 2Mb.

This ensures we get a large window on connections that transfer data, but we
don't waste a lot of the connection window on streams for which the application
is not ready to receive. Note that this only helps when streams are blocked
right from the beginning.